cast to wider unsigned type
To address `-Werror=sign-compare` warning comparing `int` with `uint32_t`

+++ src/crypto/libcrypto.c
@@ -89,7 +89,7 @@ static bool _encrypt_with_cipher(const EVP_CIPHER *cipher, aes_256_args_t args) 
         goto done;
     }
 
-    BSON_ASSERT(intermediate_bytes_written >= 0 && intermediate_bytes_written <= UINT32_MAX);
+    BSON_ASSERT(intermediate_bytes_written >= 0 && (uint64_t)intermediate_bytes_written <= UINT32_MAX);
     /* intermediate_bytes_written cannot be negative, so int -> uint32_t is OK */
     *args.bytes_written = (uint32_t)intermediate_bytes_written;
 
@@ -147,7 +147,7 @@ static bool _decrypt_with_cipher(const EVP_CIPHER *cipher, aes_256_args_t args) 
         goto done;
     }
 
-    BSON_ASSERT(intermediate_bytes_written >= 0 && intermediate_bytes_written <= UINT32_MAX);
+    BSON_ASSERT(intermediate_bytes_written >= 0 && (uint64_t)intermediate_bytes_written <= UINT32_MAX);
     /* intermediate_bytes_written cannot be negative, so int -> uint32_t is OK */
     *args.bytes_written = (uint32_t)intermediate_bytes_written;
